requirements.txt
rembg==2.0.28
Pillow==9.3.0
ffmpeg-python==0.2.0